backend/GravityFallsAPI/requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
psycopg2-binary
python-dotenv
